requirements.txt
discord.py
hcloud
humanize
pyyaml
tabulate
